Revert "Update swig wrapeprs. Don't think this actually changes anything, we just had one of the wrap files a little out of date."
Turns out that these updates were for a newer version of geocal on pistol than
what we have delivered to eco-scf. These break the older compile. No reason
right now to update eco-scf, so just back at these changes. We'll catch this
if and when we update the geocal install on eco-scf.

This reverts commit 1031ba7a6d7e9d7e70f04b434bd22f6ed4b7afdc.